2010-01-15  Carlos Alberto Cortez <calberto.cortez@gmail.com>
	* Int32.cs: In Parse() cast the parsed value to long *before* changing
	its sign - this way we preserve the original value, instead of
	accidentally changing it. This was happening parsing _exactly_
	MaxValue + 1.
	Fixes #475815.


svn path=/trunk/mcs/; revision=149646

+++ mcs/class/corlib/System/Int32.cs
@@ -586,7 +586,7 @@ namespace System {
 			
 			if (!negative && !AllowHexSpecifier){
 				if (tryParse){
-					long lval = -number;
+					long lval = -((long)number);
 
 					if (lval < MinValue || lval > MaxValue)
 						return false;
